requirements.txt
discord.py>=2.5.0
python-dotenv>=1.0.0
openai>=1.12.0
requests>=2.31.0
pydub>=0.25.1
Pillow>=10.0.0
aiohttp>=3.8.0